Working on proper mock server
