+++ crates/adapters/celestia/Cargo.toml
@@ -68,6 +68,9 @@ reltester = { workspace = true }
 sov-celestia-adapter = { path = ".", features = ["native", "arbitrary"] }
 sov-test-utils = { workspace = true, features = ["arbitrary"] }
 wiremock = "0.6"
+tonic = "0.12"
+tokio-stream = { version = "0.1", features = ["net"] }
+prost-types = "0.13"
 test-strategy = { workspace = true }
 tokio = { workspace = true, features = [
     "test-util",
@@ -104,10 +107,11 @@ bench = [
     "sov-celestia-adapter/bench",
 ]
 arbitrary = [
-    "dep:arbitrary",
-    "dep:proptest",
-    "sov-celestia-adapter/arbitrary",
-    "sov-rollup-interface/arbitrary",
-    "sov-test-utils/arbitrary",
-    "sov-modules-macros?/arbitrary"
+	"dep:arbitrary",
+	"dep:proptest",
+	"sov-celestia-adapter/arbitrary",
+	"sov-rollup-interface/arbitrary",
+	"sov-test-utils/arbitrary",
+	"sov-modules-macros?/arbitrary",
+	"prost-types/arbitrary"
 ]
